Small refactoring
Created components folder for all components that should be generalized.

Moved ChartData.js into the components folder
